vd80: Allow only one channel for analogue triggers
Triggering is global for the module. Having several channels
analog-triggering amount to trigger on the logical OR of their
trigger conditions. This is not a common case of use, and
not used at all by clients.

This patch ensures that analog triggering is being allowed just in
one channel each time.

Signed-off-by: Juan David Gonzalez Cobas <david.cobas@gmail.com>