vmebridge: add VME bus
This implements support for the VME bus, understood as in
Documentation/driver-model/bus.txt.

With this drivers can now register themselves along with the
modules they control.

In short, a VME driver calls vme_register_driver, which takes
the number of devices to be matched as a parameter. The .match
method of the driver will be called for each of these devices,
allowing the driver to discover which modules are present.

The VME bus has no auto-discovery capabilities, an illness
which the ISA bus also suffers from. Not surprisingly then, most
of the code appended has been adapted from drivers/base/isa.c,
added to the kernel in a5117ba7da.

Note that this change doesn't affect legacy drivers that
haven't yet been converted to the kernel's driver model--these
will continue working just fine.

Signed-off-by: Emilio G. Cota <cota@braap.org>