extest: do not print the prompt when invoked from a pipe
Calling the test-program through a pipe is possible. However, it is
quite painful and ugly, especially if some tool wants to parse the
output.

This removes the prompt when the test-program is called from a pipe-line.
While at it, the prompt string is NUL-terminated.

Tested on linux and lynxos.

Before the patch:

$ echo 'clk' | sis33test
sis33@cfv-864-cdv7 [00] > source:   internal
freq:   2.5e+08 Hz
[cfv-864-cdv7]$

(Note that annoyingly we start writing new input into the shell
after the prompt has been printed.)

After the patch:

$ echo 'clk' | sis33test
source: internal
freq:   2.5e+08 Hz
$
(here we start writing into the shell in a fresh and empty line)

Signed-off-by: Emilio G. Cota <cota@braap.org>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>