skel: Rawio: fix bug when two spaces have the same AM and different DW
Imagine the following XML description of address spaces of
a VME module:

<vme_space modifier="0x29" data_width="16" window_size="0x100" address="0x8000" comment=" "></vme_space>
<vme_space modifier="0x29" data_width="32" window_size="0x400" address="0x400" comment=" "></vme_space>

We want to do Rawio to the second address space, which has
data width==32.  And we can't, because Rawio will ask InsLib
to return the first address space that has 0x29 as address
modifier--i.e. we'll always do RawIO to the first space.

To fix this, the appended uses InsLibGetAddressSpaceWidth,
which also takes the datawidth as a condition to be matched
against the list of address spaces.

Signed-off-by: Emilio G. Cota <cota@braap.org>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>