skel: fix oops + hang when dealing with non-existent devices
If no valid devices are found, we free the Working Area. If
after that we try to remove the kernel module, we get a nasty
oops because we have already freed the Working Area and we
are trying to access it.

Fix this by marking the Working Area after an unsuccessful
Install() as freed setting it to null. Then Uninstall() checks
if the Wa has already been freed or not.

Signed-off-by: Emilio G. Cota <cota@braap.org>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>