utils/install: add lun--.pl
This little program takes the /etc/drivers.xml file and decrements
the logical_module_number of certain modules, printing the result
to stdout.

With this we can pipe the result to instprog as a work-around to
the feature (bug?) by which all LUNs are incremented from the
original values coming from the database:

$ lun--.pl --devices=CVORG | instprog CVORG -

The above installs all CVORG modules with LUNs as if they came
straight from the database.

Signed-off-by: Emilio G. Cota <cota@braap.org>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>