libinst: pass NULL up to the caller when something fails
When there's an error in libinst, e.g. when the parsing
fails or we run out of memory while allocating space for a
module's descriptor, we return NULL in DoElement. However,
since DoElement calls itself recursively, we do not catch
this return value from within the loop and therefore we fail
to inform the original caller of the failure.

This fixes it.

Signed-off-by: Emilio G. Cota <cota@braap.org>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>