cdcm: fix oops when a NULL address is passed to cdcm_mem_free()
When an IOCTL doesn't do any reading or writing from user-space,
such as in

 #define CVORG_IOCTLOCK		CVORG_IO  (62)

, we pass NULL to cdcm_mem_free(), because before calling the
ioctl handler we haven't allocated anything (we check there
the direction and size embedded in the IOCTL number).

The appended makes cdcm_mem_free() handle a pointer to NULL
in the same fashion as kfree() does. That means it returns
silently when the address points to NULL.

Reported-by: Gennady Sivatskiy <gennady.sivatskiy@cern.ch>
Signed-off-by: Emilio G. Cota <cota@braap.org>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>