cvorg/test: rename 'outend' command to 'starten'
In the past it made sense to call this 'output enable'. Now,
after a few development iterations, it's better to call it
'start enable', because it conveys more precisely the meaning
of this action.

Note, however, that the names of the IOCTLs and library function calls
remain untouched; it would cause too much unnecessary pain on our
users to migrate current OUT_EN and output_enable callers to enable_start.

Reported-by: Matthieu Cattin <matthieu.cattin@cern.ch>
Signed-off-by: Emilio G. Cota <cota@braap.org>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>