extest: fix infinite loop when parsing input with a minus sign preceding a non-digit
A minus sign preceding digits is parsed by extest as the beginning
of a negative number. If non-digits are found after a minus sign,
extest loops infinitely, because of a regression introduced in

   c5dc1cbc839c243703757c934cc172483d30194f
   extest: handle negative Numeric arguments
   Signed-off-by: Emilio G. Cota <cota@braap.org>

The appended fixes this by treating a minus sign preceding non-digits
as an operator, which was the existing behaviour prior to the offending
commit.

Signed-off-by: Emilio G. Cota <cota@braap.org>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>