sis3320: fix lockup when the acquisition clock freezes
When the acquisition clocks misbehaves, e.g. it's disconnected from
the device, the acquisition halts. The poor user will then call
acq_cancel, noticing that the acquisition is taking ages, and
nothing will happen--the driver will always return the module is
busy.

The problem is that the driver assumes that after writing 1 to DISARM,
the device will ALWAYS issue the 'ACQ done' interrupt. This assumption
turns out to be wrong; if the acquisition clock freezes, we may DISARM
the module 10000 times and our IRQ wouldn't yet be called.

The appended fixes this by introducing a helper (acq_complete) that
is called from both acq_cancel and inside the IRQ handler. This should
gracefully cancel any ongoing acquisition irrespective of the state of
the acquisition clock.

Reported-by: Samuel Iglesias Gonsalvez <samuel.iglesias.gonsalvez@cern.ch>
Signed-off-by: Emilio G. Cota <cota@braap.org>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>