sis33/test: added a new program for doing continuous acquisition in time.
One client requires this functionality to detect a spark in the signal, when
he does some tests.

It runs only with Python 2.7 or higher, because it imports ctypes and
multiprocessing modules. Using ctypes, it needs the shared library of libsis33
that was added in previous commit to the repository.

This program suffers problems with the bandwith limit of the DMA operation used by
the driver and delays for getting the data from the board, i.e, it may loose
some samples from the input. But it's good enough for the client, so it's implemented
in the basic way for use with external clock and tested with the lowest frequencies
available (1, 2 , 4 MHz) with relative good behaviour.

This program saves the data into text files. Gnuplot can import this files to plot.
To plot the data, should be do *by hand*. It is not our business do that inside
of the program, yet.

Signed-off-by: Samuel Iglesias Gonsalvez <samuel.iglesias.gonsalvez@cern.ch>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>