vmebus: use compat_semaphore instead of semaphore
The normal semantics of a semaphore (a variable counting the
number of free resources or pending clients waiting) is NOT
implemented in the RT patch by means of a struct semaphore, but
by means of struct comapt_semaphore, as explained in

http://www.mail-archive.com/linux-rt-users@vger.kernel.org/msg01913.html

DMA channels in vmebus are accessed through grabbing a
semaphore, so that a proc ess trying to do dma waits until
a channel becomes available or a signal arrives. Having
the standard struct semaphore caused an oops in case of
strong contention for DMA channels. Declaring dma_semaphore
as a struct compat_semaphore avoids this behavior.

Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>