cvorg/clkgen.c: added a new function clkgen_check_write
It checks that the data written to be transfered across the SPI connection was
already sent. The VHDL code has been modified for that purpose.

This solution fixes an old bug related to PLL configuration: sometimes we
change the PLL setup and a timeout appears because the operation was not
done.

Signed-off-by: Samuel Iglesias Gonsalvez <samuel.iglesias.gonsalvez@cern.ch>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>